refactor(next): relocate RootLayout to adapters/layout.tsx, checkDependencies to utilities/
