+++ packages/next/src/esbuildEntry.ts
@@ -1,4 +1,4 @@
-export { RootLayout } from './layouts/Root/index.js'
+export { RootLayout } from './adapters/layout.js'
 export { DashboardView } from './views/Dashboard/index.js'
 export { LoginView } from './views/Login/index.js'
 export { RootPage } from './views/Root/index.js'
