+++ packages/next/src/utilities/checkDependencies.ts
